Перенос actionCreator в reducer
